requirements.txt
torch>=1.8.0
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
pyyaml>=5.4.0
scipy>=1.6.0
